restructed structure, removed apps, added config.h
