+++ peripheral_drivers/i2c.c
@@ -1,13 +1,4 @@
-#define MPU6050Addr 0x68
-#define MPU6050ReadAddr 0xD1
-#define MPU6050WriteAddr 0xD0
-#define F_CPU 12000000UL
-#include<avr/io.h>
-#include<util/delay.h>
-#include<util/twi.h>
-#include<avr/interrupt.h>
-#define waitTillDone while(TWCR&(1<<TWINT)==0);
-#define SCL_CLOCK 400000
+#include "i2c.h"
 
 void TWIInit(void){
 	//Initialise the TWI hardware at 400kHz
@@ -17,118 +8,10 @@ void TWIInit(void){
 	return;
 }
 
-// void TWIStart(void){
-// 	//TWCR controls the TWI hardware
-// 	//TWSTA sends start signal.
-// 	//TWINT should be cleared by setting it.This is set at the end of the action by hardware
-// 	//TWEN enables the TWI
-// 	TWCR = (1<<TWINT)|(1<<TWSTA)|(1<<TWEN);
-// 	waitTillDone;
-// }
-//
-// void TWIStop(void){
-// 	//TWSTO sends stop signal
-// 	TWCR = (1<<TWINT)|(1<<TWSTO)|(1<<TWEN);
-// }
-//
-// void TWIWrite(uint8_t data){
-// 	//This sends data level 1  abstraction.
-// 	//TWDR holds data to be written
-// 	TWDR = data;
-// 	TWCR = (1<<TWINT)|(1<<TWEN);
-// 	waitTillDone;
-// }
-//
-// uint8_t TWIReadACK(void){
-// 	//Read data and acknowledge
-// 	uint8_t x;
-// 	x=TWDR;
-// 	TWCR = (1<<TWINT)|(1<<TWEN)|(1<<TWEA);
-// 	waitTillDone;
-// 	_delay_us(50);
-// 	return TWDR;
-// }
-//
-// uint8_t TWIReadNACK(void){
-// 	uint8_t x;
-// 	x=TWDR;
-// 	//Read data and acknowledge
-// 	TWCR = ((1<<TWINT)|(1<<TWEN));
-// 	waitTillDone;
-// 	_delay_us(50);
-// 	return TWDR;
-// }
-//
-// uint8_t TWIGetStatus(void){
-// 	uint8_t status;
-// 	status = TWSR&0xf8;
-// 	return status;
-//
-// }
-//
-// void TWIWriteReg(uint8_t reg, uint8_t data){
-//         TWIStart();
-//
-//         _delay_us(10);
-//         while(TWIGetStatus()!=0x08);
-//         TWIWrite(MPU6050WriteAddr);
-//
-//         _delay_us(10);
-//         while(TWIGetStatus()!=0x18);
-//         TWIWrite(reg);
-//         _delay_us(10);
-//
-//         while(TWIGetStatus()!=0x28);
-//         TWIWrite(data);
-//
-//         while(TWIGetStatus()!=0x28);
-//         TWIStop();
-//         return;
-//
-// }
-//
-//
-// uint8_t TWIReadReg(uint8_t reg){
-// 	uint8_t regdata;
-// 	TWIStart();
-//
-//         _delay_us(10);
-//         while(TWIGetStatus()!=0x08);
-//         TWIWrite(MPU6050WriteAddr);
-//
-//         _delay_us(10);
-//         while(TWIGetStatus()!=0x18);
-//         TWIWrite(reg);
-//         _delay_us(10);
-//
-//         while(TWIGetStatus()!=0x28);
-//         TWIStart();
-//         _delay_us(10);
-//
-//         while(TWIGetStatus()!=0x10);
-//         TWIWrite(MPU6050ReadAddr);
-//         _delay_us(10);
-//
-//         while(TWIGetStatus()!=0x40);
-//
-//         regdata = TWIReadNACK();
-//          _delay_us(10);
-//         while(TWIGetStatus()!=0x58);
-//
-//         TWIStop();
-//
-//         return regdata;
-//
-//
-// }
-//
-
-
-
 // from peter flury
 
 unsigned char i2c_start(unsigned char address)
-{   
+{
     uint8_t   twst;
 
 	// send START condition
@@ -197,11 +80,12 @@ unsigned char i2c_readNak(void)
 	cli();
 	TWCR = (1<<TWINT) | (1<<TWEN);
 	while(!(TWCR & (1<<TWINT)));
-sei();
+	sei();
     return TWDR;
 
 
-}/* i2c_readNak */
+}
+/* i2c_readNak */
 
 
 void i2c_stop(void)
@@ -219,8 +103,8 @@ void i2c_stop(void)
 void i2c_init(void)
 {
   /* initialize TWI clock: 100 kHz clock, TWPS = 0 => prescaler = 1 */
-cli();
+  cli();
   TWSR = 0;                         /* no prescaler */
   TWBR = ((F_CPU/SCL_CLOCK)-16)/2;  /* must be > 10 for stable operation */
-sei();
+  sei();
 }/* i2c_init */
