+++ utils.h
@@ -1,0 +1,8 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include<avr/io.h>
+
+void strcut(char* , char* ,uint8_t ,uint8_t);
+
+#endif
